% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deSolve_pecc.R
\name{deSolve_pecc}
\alias{deSolve_pecc}
\title{deSolve_pecc}
\usage{
deSolve_pecc(model)
}
\arguments{
\item{model}{The model written by the user}
}
\description{
Analyze simplified deSolve syntax to extract all informations
}
