% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desolve_to_monolix.R
\name{ode_monolix}
\alias{ode_monolix}
\title{ODE monolix}
\usage{
ode_monolix(
  func = Lorenz,
  mb_output,
  parms = parameters,
  y = state,
  depot = "C",
  datafile = "",
  add_param = "",
  ytype = data.frame(),
  outputcat = T,
  ...
)
}
\arguments{
\item{func}{same as the ode}

\item{parms}{same as the ode}

\item{y}{same as the ode}

\item{datafile}{Either a .tab or an other cfl}

\item{...}{}
}
\value{

}
\description{
ODE monolix
}
