% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desolve_to_nonmem.R
\name{ode_nonmem}
\alias{ode_nonmem}
\title{DeSolve pharmacometrics model creator}
\usage{
ode_nonmem(
  func = Lorenz,
  omega,
  parms = parameters,
  y = state,
  mb_output,
  add_param = "",
  mu_referencig = T,
  BLQ = F,
  datafile = "",
  outputcat = T,
  ...
)
}
\arguments{
\item{func}{Function (same as deSolve in functions)}

\item{parms}{Parameters (same as deSolve in functions)}

\item{y}{same as the ode}

\item{add_param}{vector of parameter names to add, mostly usefull for adding baseline parameters.
If it has the format "X1_0", with  X1 being the name of a cmt, add it as a baseline for the concerned cmt.
If the format is "Z = 21", add a new parameter with the name Z and initial value 21.}

\item{mu_referencig}{booléen, declaration of parameters with/without mu referencing.}

\item{datafile}{Either a .tab or an other cfl}

\item{output}{vector of the output names. In case of several outputs, use the format c("name1 = X", "name2 = Y",...), with X/Y being the YTYPE number.}
}
\value{

}
\description{
Create a NONMEM code with a deSolve model.
}
\author{
Thibaud Derippe
}
